+++ securedrop/source_templates/error.html
@@ -4,5 +4,5 @@
 
 <p>{{ gettext('Sorry, the website encountered an error and was unable to complete your request.') }}</p>
 
-<p><a href="/login">{{ gettext('Look up a codename...') }}</a></p>
+<p><a href="{{ url_for('main.login') }}">{{ gettext('Look up a codename...') }}</a></p>
 {% endblock %}
