+++ securedrop/source_templates/login.html
@@ -5,7 +5,7 @@
 
 {% include 'flashed.html' %}
 
-<form method="post" action="/login" autocomplete="off">
+<form method="post" action="{{ url_for('main.login') }}" autocomplete="off">
 <input name="csrf_token" type="hidden" value="{{ csrf_token() }}">
 
 <p class="center"><input id="login-with-existing-codename" type="password" name="codename" class="codename-box" autocomplete="off" placeholder="{{ gettext('Enter your codename') }}" autofocus></p>
