+++ securedrop/source_templates/notfound.html
@@ -4,5 +4,5 @@
 
 <p id="page-not-found">{{ gettext("Sorry, we couldn't locate what you requested.") }}</p>
 
-<p><a href="/login">{{ gettext('Look up a codename...') }}</a></p>
+<p><a href="{{ url_for('main.login') }}">{{ gettext('Look up a codename...') }}</a></p>
 {% endblock %}
